// types/proto/reminder.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.7
//   protoc               v6.32.0
// source: proto/reminder.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from '@nestjs/microservices';
import { Observable } from 'rxjs';

export const protobufPackage = 'reminder';

export interface CreateReminderRequest {
  userId: string;
  title: string;
  description: string;
  /** ISO 8601 string */
  scheduledAt: string;
}

export interface ReminderResponse {
  id: string;
  message: string;
}

export const REMINDER_PACKAGE_NAME = 'reminder';

export interface ReminderServiceClient {
  createReminder(request: CreateReminderRequest): Observable<ReminderResponse>;
}

export interface ReminderServiceController {
  createReminder(
    request: CreateReminderRequest
  ):
    | Promise<ReminderResponse>
    | Observable<ReminderResponse>
    | ReminderResponse;
}

export function ReminderServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ['createReminder'];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method
      );
      GrpcMethod('ReminderService', method)(
        constructor.prototype[method],
        method,
        descriptor
      );
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method
      );
      GrpcStreamMethod('ReminderService', method)(
        constructor.prototype[method],
        method,
        descriptor
      );
    }
  };
}

export const REMINDER_SERVICE_NAME = 'ReminderService';